Latest vectorstorm, build process updated.
Moved OSX-specific files out of the project root.

Added a process to build a legal, relocatable app bundle on OS X.  Use
the 'install' target to get that.  (This copies the required
non-standard libraries into the bundle, and sets their rpath correctly
so that the app bundle can be copied to another machine).  I've done
this process manually for almost every OSX release I've ever done --
finally I have a process which will do it automatically as part of the
build.